#77: rename UserAvatarWall to UsersAwatarWall & add test cases
