+++ src/features/usersSearch/view/components/UsersAvatarsWall/UsersAvatarsWall.scss
